Feature/reviewlogic

+++ PCBuilder.Services.CustomerAPI/Models/Order.cs
@@ -8,7 +8,7 @@ public class Order
     public Guid? UserId { get; set; }
     public int ReviewId { get; set; }
     public int? ComputerId { get; set; }
-    public int Budget { get; set; }
+    public decimal Budget { get; set; }
     public decimal SellingPrice { get; set; }
     public string Description { get; set; } = string.Empty;
     public string DetailedDescription { get; set; } = string.Empty;
